refactor(cargo-wdk): move action types into the modules that own them
Relocate shared types out of `actions/mod.rs` so each lives in the action
module that uses it, leaving `actions/mod.rs` as a pure module-tree root:

- `Profile`, `to_target_triple` (+ target-triple consts) -> `actions/build`
- `DriverType`, `KMDF_STR`/`UMDF_STR`/`WDM_STR` -> `actions/new`

Update imports in `cli.rs`, `build/mod.rs`, `build/build_task.rs`,
`build/tests.rs`, and `new/mod.rs` accordingly. Matches the existing
placement of `SignMode`/`TargetPlatform` in the `build` module.
