+++ crates/cargo-wdk/src/actions/mod.rs
@@ -10,85 +10,3 @@
 pub mod build;
 pub mod clean;
 pub mod new;
-
-use std::{
-    fmt::{self, Display},
-    str::FromStr,
-};
-
-use wdk_build::CpuArchitecture;
-
-pub const KMDF_STR: &str = "kmdf";
-pub const UMDF_STR: &str = "umdf";
-pub const WDM_STR: &str = "wdm";
-/// `x86_64/Amd64` target triple name
-const X86_64_TARGET_TRIPLE_NAME: &str = "x86_64-pc-windows-msvc";
-/// `aarch64/Arm64` target triple name
-const AARCH64_TARGET_TRIPLE_NAME: &str = "aarch64-pc-windows-msvc";
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Profile {
-    Dev,
-    Release,
-}
-impl FromStr for Profile {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "dev" => std::result::Result::Ok(Self::Dev),
-            "release" => std::result::Result::Ok(Self::Release),
-            _ => Err(format!("'{s}' is not a valid profile")),
-        }
-    }
-}
-impl Display for Profile {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let s = match self {
-            Self::Dev => "dev",
-            Self::Release => "release",
-        };
-        write!(f, "{s}")
-    }
-}
-
-/// Converts `CpuArchitecture` to its corresponding target triple name.
-#[must_use]
-pub fn to_target_triple(cpu_arch: CpuArchitecture) -> String {
-    match cpu_arch {
-        CpuArchitecture::Amd64 => X86_64_TARGET_TRIPLE_NAME.to_string(),
-        CpuArchitecture::Arm64 => AARCH64_TARGET_TRIPLE_NAME.to_string(),
-    }
-}
-
-/// Enum of driver types.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum DriverType {
-    Kmdf,
-    Umdf,
-    Wdm,
-}
-
-impl FromStr for DriverType {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            KMDF_STR => Ok(Self::Kmdf),
-            UMDF_STR => Ok(Self::Umdf),
-            WDM_STR => Ok(Self::Wdm),
-            _ => Err(format!("'{s}' is not a valid driver type")),
-        }
-    }
-}
-
-impl Display for DriverType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let s = match self {
-            Self::Kmdf => KMDF_STR,
-            Self::Umdf => UMDF_STR,
-            Self::Wdm => WDM_STR,
-        };
-        write!(f, "{s}")
-    }
-}
